reuse useNodePB (useNodeData)
