+++ src/javascript/JContent/EditFrame/Box/Box.jsx
@@ -5,12 +5,12 @@ import styles from './Box.scss';
 import {useBoxStatus} from './useBoxStatus';
 import {useNodeDragPB} from '~/JContent/dnd/useNodeDragPB';
 import editStyles from '../EditFrame.scss';
-import {useNodeDropPB} from '~/JContent/dnd/useNodeDropPB';
 import {DefaultBar} from '~/JContent/EditFrame/DefaultBar';
 import {getBoundingBox} from '~/JContent/EditFrame/EditFrame.utils';
 import {Breadcrumbs} from '../Breadcrumbs';
 import {findAvailableBoxConfig, hasMixin} from '../../JContent.utils';
 import {useTranslation} from 'react-i18next';
+import {useNodeDropData} from '~/JContent/dnd';
 
 const reposition = function (element, currentOffset, setCurrentOffset, isHeaderDisplayed) {
     const box = getBoundingBox(element, isHeaderDisplayed);
@@ -139,7 +139,7 @@ export const Box = React.memo(({
 
     const rootDiv = useRef();
 
-    const [{isCanDrop, insertPosition, destParent, isOver}, drop] = useNodeDropPB({
+    const [{isCanDrop, insertPosition, destParent, isOver}, drop] = useNodeDropData({
         dropTarget: node,
         orderable: Boolean(parent),
         entries,
