+++ src/javascript/JContent/dnd/index.js
@@ -1,5 +1,5 @@
 export * from './DragLayer';
 export * from './useFileDrop';
-export * from './useNodeDrop';
+export {useNodeDropData, useNodeDrop} from './useNodeDrop';
 export * from './useNodeDrag';
 export {useConnector} from '~/JContent/dnd/useConnector';
